Modify index-create to allow simple indexes
Simple indexes are based on the value of a single field, and can be
created by passing the field name as the name of the index.
N.B. You can create an index on a field that doesn't 'exist', because
RethinkDB has no notion of a schema and therefore no ability to tell
whether a field will be present in the future or an index is 'valid'
when it is created.

Fixes #83